# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateWebhookRequest:
    properties:
      active:
        example: true
        type: boolean
      events:
        example:
        - user.created
        - user.updated
        items:
          type: string
        type: array
      url:
        example: https://example.com/webhook
        type: string
    required:
    - events
    - url
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: securepassword123
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: newpassword123
        type: string
      username:
        example: john_doe
        type: string
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateWebhookRequest:
    properties:
      active:
        example: false
        type: boolean
      events:
        example:
        - user.created
        items:
          type: string
        type: array
      url:
        example: https://example.com/webhook
        type: string
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_user.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_user.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Sovrabase API server.
  title: Sovrabase API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.LoginResponse'
      summary: Connexion API Admin
      tags:
      - User
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Logout
      tags:
      - User
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.LoginResponse'
      summary: Refresh Token
      tags:
      - User
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.LoginResponse'
      summary: Enregistrement d'utilisateur
      tags:
      - User
...
  /project/{id}/webhooks:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Webhooks
      tags:
      - Webhook
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Webhook
      tags:
      - Webhook
  /project/{id}/webhooks/{webhook_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Webhook
      tags:
      - Webhook
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Webhook
      tags:
      - Webhook
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      - description: Webhook update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update Webhook
      tags:
      - Webhook
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.User'
      security:
      - Bearer: []
      summary: Get User
      tags:
      - User
    patch:
      consumes:
      - application/json
      parameters:
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_user.User'
      security:
      - Bearer: []
      summary: Update User
      tags:
      - User
swagger: "2.0"