# internal/api/handlers/user.go
package handlers

import (
	"encoding/json"
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models/requests"
	"github.com/ketsuna-org/sovrabase/internal/models/user"
)

// LoginHandler handles user login
// @Summary Connexion API Admin
// @Tags User
// @Accept json
// @Produce json
// @Param request body user.LoginRequest true "Login credentials"
// @Success 200 {object} user.LoginResponse
// @Router /auth/login [post]
func LoginHandler(w http.ResponseWriter, r *http.Request) {
	var req user.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement actual authentication logic
	// For now, mock response
	response := user.LoginResponse{
		AccessToken:  "mock_access_token",
		RefreshToken: "mock_refresh_token",
		TokenType:    "Bearer",
		ExpiresIn:    3600,
		Scope:        "read write",
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// RefreshTokenHandler refreshes the access token
// @Summary Refresh Token
// @Tags User
// @Accept json
// @Produce json
// @Param request body requests.RefreshTokenRequest true "Refresh token"
// @Success 200 {object} user.LoginResponse
// @Router /auth/refresh [post]
func RefreshTokenHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.RefreshTokenRequest
	_ = req
	// TODO: Implement token refresh logic
	response := user.LoginResponse{
		AccessToken:  "mock_new_access_token",
		RefreshToken: "mock_new_refresh_token",
		TokenType:    "Bearer",
		ExpiresIn:    3600,
		Scope:        "read write",
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// GetUserHandler gets the current user's information
// @Summary Get User
// @Tags User
// @Security Bearer
// @Produce json
// @Success 200 {object} user.User
// @Router /user [get]
func GetUserHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get user logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"id":       "user1",
		"username": "testuser",
		"email":    "test@example.com",
	})
}

// UpdateUserHandler updates the current user's information
// @Summary Update User
// @Tags User
// @Security Bearer
// @Accept json
// @Produce json
// @Param request body requests.UpdateUserRequest true "User update data"
// @Success 200 {object} user.User
// @Router /user [patch]
func UpdateUserHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateUserRequest
	_ = req
	// TODO: Implement update user logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"id":       "user1",
		"username": "updateduser",
		"email":    "updated@example.com",
	})
}

// LogoutHandler logs out the current user
// @Summary Logout
// @Tags User
// @Security Bearer
// @Produce json
// @Success 200 {object} map[string]string
// @Router /auth/logout [post]
func LogoutHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement logout logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"message": "Successfully logged out",
	})
}

// RegisterHandler registers a new admin user
// @Summary Enregistrement d'utilisateur
// @Tags User
// @Accept json
// @Produce json
// @Param request body requests.RegisterRequest true "Registration data"
// @Success 200 {object} user.LoginResponse
// @Router /auth/register [post]
func RegisterHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.RegisterRequest
	_ = req
	// TODO: Implement registration logic
	response := user.LoginResponse{
		AccessToken:  "mock_access_token",
		RefreshToken: "mock_refresh_token",
		TokenType:    "Bearer",
		ExpiresIn:    3600,
		Scope:        "read write",
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}