# internal/api/handlers/webhook.go
package handlers

import (
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models/requests"
)

// GetWebhooksHandler gets all webhooks for a project
// @Summary Get Webhooks
// @Tags Webhook
// @Security Bearer
// @Param id path string true "Project ID"
// @Success 200
// @Router /project/{id}/webhooks [get]
func GetWebhooksHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get webhooks logic
	w.WriteHeader(http.StatusOK)
}

// CreateWebhookHandler creates a new webhook
// @Summary Create Webhook
// @Tags Webhook
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param request body requests.CreateWebhookRequest true "Webhook creation data"
// @Success 200
// @Router /project/{id}/webhooks [post]
func CreateWebhookHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.CreateWebhookRequest
	_ = req
	// TODO: Implement create webhook logic
	w.WriteHeader(http.StatusOK)
}

// GetWebhookHandler gets a specific webhook
// @Summary Get Webhook
// @Tags Webhook
// @Security Bearer
// @Param id path string true "Project ID"
// @Param webhook_id path string true "Webhook ID"
// @Success 200
// @Router /project/{id}/webhooks/{webhook_id} [get]
func GetWebhookHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get webhook logic
	w.WriteHeader(http.StatusOK)
}

// UpdateWebhookHandler updates a webhook
// @Summary Update Webhook
// @Tags Webhook
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param webhook_id path string true "Webhook ID"
// @Param request body requests.UpdateWebhookRequest true "Webhook update data"
// @Success 200
// @Router /project/{id}/webhooks/{webhook_id} [patch]
func UpdateWebhookHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateWebhookRequest
	_ = req
	// TODO: Implement update webhook logic
	w.WriteHeader(http.StatusOK)
}

// DeleteWebhookHandler deletes a webhook
// @Summary Delete Webhook
// @Tags Webhook
// @Security Bearer
// @Param id path string true "Project ID"
// @Param webhook_id path string true "Webhook ID"
// @Success 200
// @Router /project/{id}/webhooks/{webhook_id} [delete]
func DeleteWebhookHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete webhook logic
	w.WriteHeader(http.StatusOK)
}